# AutoriaFinal-Frontend/apidoc.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Autoria Auction API",
    "description": "API for auction management system",
    "version": "v1"
  },
  "paths": {
    "/api/Auction": {
      "get": {
        "tags": [
          "Auction"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuctionGetDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuctionGetDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuctionGetDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Auction"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuctionCreateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AuctionCreateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AuctionCreateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDetailDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Auction/{id}": {
      "get": {
        "tags": [
          "Auction"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDetailDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Auction"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuctionUpdateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AuctionUpdateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AuctionUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDetailDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Auction"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Auction/{id}/start": {
      "post": {
        "tags": [
          "Auction"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDetailDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Auction/{id}/end": {
      "post": {
        "tags": [
          "Auction"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Auction/{id}/cancel": {
      "post": {
        "tags": [
          "Auction"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuctionCancelRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AuctionCancelRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AuctionCancelRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Auction/{id}/extend": {
      "post": {
        "tags": [
          "Auction"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuctionExtendRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AuctionExtendRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AuctionExtendRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Auction/{id}/next-car": {
      "post": {
        "tags": [
          "Auction"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDetailDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Auction/{id}/set-current-car": {
      "post": {
        "tags": [
          "Auction"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetCurrentCarRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SetCurrentCarRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SetCurrentCarRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDetailDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Auction/location/{locationId}": {
      "get": {
        "tags": [
          "Auction"
        ],
        "parameters": [
          {
            "name": "locationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuctionGetDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuctionGetDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AuctionGetDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Auction/{id}/timer": {
      "get": {
        "tags": [
          "Auction"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionTimerInfo"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionTimerInfo"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionTimerInfo"
                }
              }
            }
          }
        }
      }
    },
    "/api/Auction/{id}/current-state": {
      "get": {
        "tags": [
          "Auction"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDetailDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDetailDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuctionDetailDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuctionGetDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "startTimeUtc": {
            "type": "string",
            "format": "date-time"
          },
          "endTimeUtc": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "startPrice": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "isLive": {
            "type": "boolean"
          },
          "currentCarLotNumber": {
            "type": "string",
            "nullable": true
          },
          "totalCarsCount": {
            "type": "integer",
            "format": "int32"
          },
          "carsWithPreBidsCount": {
            "type": "integer",
            "format": "int32"
          },
          "locationId": {
            "type": "string",
            "format": "uuid"
          },
          "locationName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "AuctionUpdateDto": {
        "type": "object",
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme.",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ]
}